license: use CC BY instead of BY-SA.
The ShareAlike variant could be interpreted as covering data generated
from our data (much like the LieDB web site), which is not our intent.
We only want to require people to share changes made to the actual
data.  Unfortunately I don't know how to make that distinction in a
license, hence simply dropping the SA requirement.  Hopefully people
will still do the right thing, if only to reduce the maintentance
burden on themselves…