Add option to build Dart version.
There isn't actually a Dart version in master yet, but this will reduce
the diff between master and wip/dart so it's easier to rebase
periodically.

Having the option toggleable in config.toml will also make it a bit
less inconvenient to test the static site once the dart version lands
in master.