Significant restructuring of the site.
I still need to port the API over, but the HTML interface should work.

The major change is that I've split individual assertions (which I
plan to start calling "statements" instead of "assertions") into
separate pages nested under each person, and am using a taxonomy to
manage claims (I plan to also use taxonomies for groups and themes).

The prime motivator for this was to make it possible to have a 1:N
relationship for statements to claims, but I'm making a lot of other
changes at the same time since I have to rewrite so much stuff anyways.

This also includes some new claims that popped up while I was working
on this change.