Pass page as well as statement to quote partial.
This will let us do stuff conditionally in the quote partial without
having to duplicate everything. For example, the person's name
doesn't really need to be a link when we're already on that page, and
when on a claim page we don't really need to include it in the list of
claims.