Move fact checks from claims to individual statements.
Part of this is that fact checks can apply to multiple claims, or
a single claim of a multi-claim statement. The drawback is that
fact checks of multiple statements must be duplicated for each
statement. The work-around for this would be adding a fact check
type which maps to both claims and statements, but I think the
usability drawbacks would be too great to justify the deduplication.

This does also add a few fact checks I came across while working on
moving the data.