Clean up us-highest-business-tax-rate description a little bit.
There aren't really any substantive changes, just cleaning up the prose
a little.

The biggest change is that "business" in the title has been changed to
"corporate". Both variations are addressed in the description, but I'd
rather be more generous in the title than present an opportunity for
people to claim we're mis-interpreting their words.