Reduce Tweet count for multiple statments of the same claim in 1 commit
Before this, every statement would generate a broadcast Tweet (and,
if the statement was made on Twitter, a reply tweet). Now multiple
statements of the same claim by the same person will only generate a
single tweet (plus replies to any tweets if it's a statement made on
twitter). This should generate a lot less noise when going through
archives to find examples of someone stating the same lie multiple
times.